feat: App 컴포넌트에 SafeAreaView 래퍼 추가
