+++ App.tsx
@@ -15,6 +15,7 @@ import OppuScreen from './src/screens/OppuScreen';
 import AddScreen from './src/screens/AddScreen';
 import ArchiveScreen from './src/screens/ArchiveScreen';
 import MyScreen from './src/screens/MyScreen';
+import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
 
 const Tab = createBottomTabNavigator();
 
@@ -22,20 +23,17 @@ function App() {
   const isDarkMode = useColorScheme() === 'dark';
 
   return (
-    <View style={styles.container}>
-      <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
-      <>
-        <NavigationContainer>
-          <Tab.Navigator tabBar={props => <BottomTabBar {...props} />}>
-            <Tab.Screen name="HOME" component={HomeScreen} />
-            <Tab.Screen name="OPPU" component={OppuScreen} />
-            <Tab.Screen name="ADD" component={AddScreen} />
-            <Tab.Screen name="ARCHIVE" component={ArchiveScreen} />
-            <Tab.Screen name="MY" component={MyScreen} />
-          </Tab.Navigator>
-        </NavigationContainer>
-      </>
-    </View>
+    <SafeAreaView style={styles.container}>
+      <NavigationContainer>
+        <Tab.Navigator tabBar={props => <BottomTabBar {...props} />}>
+          <Tab.Screen name="HOME" component={HomeScreen} />
+          <Tab.Screen name="OPPU" component={OppuScreen} />
+          <Tab.Screen name="ADD" component={AddScreen} />
+          <Tab.Screen name="ARCHIVE" component={ArchiveScreen} />
+          <Tab.Screen name="MY" component={MyScreen} />
+        </Tab.Navigator>
+      </NavigationContainer>
+    </SafeAreaView>
   );
 }
