+++ src/components/common/BottomTabBar/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { View, TouchableOpacity, Text, StyleSheet } from 'react-native';
 import type { BottomTabBarProps } from '@react-navigation/bottom-tabs';
+import styles from './styles';
 //import FeatherIcon from 'react-native-vector-icons/Feather';
 
 const BottomTabBar: React.FC<BottomTabBarProps> = ({ state, navigation }) => {
@@ -60,43 +61,5 @@ const BottomTabBar: React.FC<BottomTabBarProps> = ({ state, navigation }) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    paddingVertical: 12,
-    paddingHorizontal: 16,
-  },
-  tab: {
-    flex: 1,
-    flexDirection: 'column',
-    gap: 4,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  icon: {
-    fontSize: 24,
-    color: '#A3A9B6',
-  },
-  addCircle: {
-    backgroundColor: '#34618E',
-    borderRadius: 100,
-    width: 56,
-    height: 56,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  addText: {
-    color: '#fff',
-    fontSize: 32,
-    includeFontPadding: false,
-    textAlign: 'center',
-    textAlignVertical: 'center',
-  },
-  label: {
-    color: '#6B7280',
-    fontSize: 14,
-  },
-});
 
 export default BottomTabBar;
